requirements.txt
# production dependencies
anyio==4.6.2.post1
asgiref==3.8.1
attrs==24.2.0
certifi==2024.8.30
Django==5.1.2
djangorestframework==3.15.2
drf-spectacular==0.27.2
gunicorn==23.0.0
h11==0.14.0
hiredis==3.0.0
httpcore==1.0.6
httpx==0.27.2
idna==3.10
inflection==0.5.1
jsonschema==4.23.0
jsonschema-specifications==2024.10.1
packaging==24.1
psycopg==3.2.3
psycopg-binary==3.2.3
PyYAML==6.0.2
redis==5.2.0
referencing==0.35.1
rpds-py==0.21.0
sniffio==1.3.1
sqlparse==0.5.1
typing_extensions==4.12.2
uritemplate==4.1.1

# Typing
mypy==1.11.2
mypy-extensions==1.0.0
django-stubs==5.1.0
django-stubs-ext==5.1.1
djangorestframework-stubs==3.15.1
typing_extensions==4.12.2


# Development
ruff==0.7.2
ipython==8.29.0
coverage==7.6.4